// src/api/generated/model/apiResponseListBranchDto.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { BranchDto } from "./branchDto";

export interface ApiResponseListBranchDto {
  success?: boolean;
  message?: string;
  data?: BranchDto[];
  errors?: string[];
  timestamp?: string;
  path?: string;
}