// src/api/generated/model/updateExamRequest.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateExamRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  examTypeId: string;
  /** @minItems 1 */
  questionGroupIds: string[];
  branchId: string;
  brandId: string;
}